Move scala2/scala3 apis to separate files
